requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
Pillow
